added create/update event, create/update eventcategory endpoints and services

+++ backend/ticket-management-system/src/main/java/com/masterticket/platform/ticketmanagementsystem/Repo/BookingRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masterticket.platform.ticketmanagementsystem.Models.Booking;
 
-public interface BookingRepo extends JpaRepository<Booking, Long> {
+public interface BookingRepo extends JpaRepository<Booking, Integer> {
     
 }
