+++ backend/ticket-management-system/src/main/java/com/masterticket/platform/ticketmanagementsystem/Repo/EventCategoryRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masterticket.platform.ticketmanagementsystem.Models.EventCategory;
 
-public interface EventCategoryRepo extends JpaRepository<EventCategory, Long> {
+public interface EventCategoryRepo extends JpaRepository<EventCategory, Integer> {
     
 }
