+++ backend/ticket-management-system/src/main/java/com/masterticket/platform/ticketmanagementsystem/Repo/EventRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masterticket.platform.ticketmanagementsystem.Models.Event;
 
-public interface EventRepo extends JpaRepository<Event, Long> {
+public interface EventRepo extends JpaRepository<Event, Integer> {
     
 }
