+++ backend/ticket-management-system/src/main/java/com/masterticket/platform/ticketmanagementsystem/Repo/IssuedTicketRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masterticket.platform.ticketmanagementsystem.Models.IssuedTicket;
 
-public interface IssuedTicketRepo extends JpaRepository<IssuedTicket, Long> {
+public interface IssuedTicketRepo extends JpaRepository<IssuedTicket, Integer> {
     
 }
