+++ backend/ticket-management-system/src/main/java/com/masterticket/platform/ticketmanagementsystem/Repo/TicketingOfficerRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masterticket.platform.ticketmanagementsystem.Models.TicketingOfficer;
 
-public interface TicketingOfficerRepo extends JpaRepository<TicketingOfficer, Long> {
+public interface TicketingOfficerRepo extends JpaRepository<TicketingOfficer, Integer> {
     
 }
